Add files via upload. Data files too big
Ok so I added a few logical tests like 4 of them. I DO NOT HAVE FILTER FOR MANHATTAN. I also have more columns for visualtions which I will add that notebook in awhile just the stress tests he asked for. Than I checked for duplicates and dropped all NANs as well. I added seasons, day of the week, holidays and everything as well so this is all just for analysing and maybe for the model. Logical test 5 was working and now is being rude and not working. So checking logic still for this its all in caps. Nothing is done with last logic test but yeah everything broke down at the end