adding files broken up from first one
I have them all in the order they need to be run to get the files. I split up the file from earlier. The third one is visualisations of different things to be run. it is not filtered to manhattan though so will need to be changed but just earlier look into things and barcharts and broken down. i will expand it later. Here it is